Move config tests in a separate file
